apps/ml/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
asyncpg==0.30.*
pandas==2.2.*
numpy>=1.26,<3
scikit-learn==1.6.*
xgboost==2.1.*
joblib==1.4.*
ta==0.11.*
python-dotenv==1.0.*
torch>=2.2,<3
